[dashboard] (refactor) extract UsageView to be reused for individual
